moved util file
